- done Game Play

+++ Assets/_Game/Scripts/Configs/EnemyConfig.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class CharacterData : BaseData
 {
     public EnemyType Type;
-    public float Health = 0;
+    public int Health = 0;
     public float Speed = 0;
     public int Damage = 0;
     public GameObject Prefab;
@@ -16,7 +16,7 @@ public class CharacterInfo
 {
     public int CharId;
     public int Level;
-    public float Health = 0;
+    public int Health = 0;
     public float Speed = 0;
     public int Damage = 0;
 }
