+++ Assets/_Game/Scripts/Editor/Tab/EnemyTab.cs
@@ -21,7 +21,7 @@ public class EnemyTab : TabContent
                 id = GameConfig.Ins.EnemyDatas.Count > 0 ? GameConfig.Ins.EnemyDatas[^1].id + 1 : 1,
                 name = $"New Enemy{(GameConfig.Ins.EnemyDatas.Count > 0 ? GameConfig.Ins.EnemyDatas[^1].id + 1 : 1)}",
                 Type = EnemyType.Normal, // Default value, adjust as needed
-                Health = 10f,
+                Health = 10,
                 Speed = 2f,
                 Damage = 1,
                 Prefab = null,
@@ -38,7 +38,7 @@ public class EnemyTab : TabContent
         enemyTable.AddCol("ID", 50, e => { e.id = Draw.Int(e.id, 30); Draw.Space(20); });
         enemyTable.AddCol("Name", 120, e => { e.name = Draw.Text(e.name, 100); Draw.Space(20); });
         enemyTable.AddCol("Type", 100, e => { e.Type = (EnemyType)Draw.Enum(e.Type, 80); Draw.Space(20); });
-        enemyTable.AddCol("Health", 80, e => { e.Health = Draw.Float(e.Health, 60); Draw.Space(20); });
+        enemyTable.AddCol("Health", 80, e => { e.Health = Draw.Int(e.Health, 60); Draw.Space(20); });
         enemyTable.AddCol("Speed", 80, e => { e.Speed = Draw.Float(e.Speed, 60); Draw.Space(20); });
         enemyTable.AddCol("Damage", 80, e => { e.Damage = Draw.Int(e.Damage, 60); Draw.Space(20); });
         enemyTable.AddCol("Prefab", 150, e => { e.Prefab = Draw.Object(e.Prefab, 130); Draw.Space(20); });
